refactor: move prompt method to type signer
