--- src/types/icrc-prompts.ts
@@ -1,14 +1,0 @@
-import {z} from 'zod';
-import {
-  ICRC25_REQUEST_PERMISSIONS,
-  ICRC27_ACCOUNTS,
-  ICRC49_CALL_CANISTER
-} from '../constants/icrc.constants';
-
-export const IcrcApproveMethodSchema = z.enum([
-  ICRC25_REQUEST_PERMISSIONS,
-  ICRC27_ACCOUNTS,
-  ICRC49_CALL_CANISTER
-]);
-
-export type IcrcApproveMethod = z.infer<typeof IcrcApproveMethodSchema>;
