+++ src/types/signer-prompts.ts
@@ -1,7 +1,20 @@
 import {z} from 'zod';
+import {
+  ICRC25_REQUEST_PERMISSIONS,
+  ICRC27_ACCOUNTS,
+  ICRC49_CALL_CANISTER
+} from '../constants/icrc.constants';
 import type {icrc21_consent_info} from '../declarations/icrc-21';
 import {IcrcAccountsSchema} from './icrc-accounts';
 import {IcrcScopesArraySchema} from './icrc-responses';
+
+export const IcrcApproveMethodSchema = z.enum([
+  ICRC25_REQUEST_PERMISSIONS,
+  ICRC27_ACCOUNTS,
+  ICRC49_CALL_CANISTER
+]);
+
+export type IcrcApproveMethod = z.infer<typeof IcrcApproveMethodSchema>;
 
 const PermissionsConfirmationSchema = z.function().args(IcrcScopesArraySchema).returns(z.void());
